+++ include/vrv/functorparams.h
@@ -1896,7 +1896,8 @@ public:
 
 /**
  * member 0: the current Syl
- * member 1: the last Note
+ * member 1: the last Note or Chord
+ * member 2: the penultimate Note or Chord
  **/
 
 class PrepareLyricsParams : public FunctorParams {
@@ -1904,12 +1905,12 @@ public:
     PrepareLyricsParams()
     {
         m_currentSyl = NULL;
-        m_lastNote = NULL;
-        m_lastButOneNote = NULL;
+        m_lastNoteOrChord = NULL;
+        m_penultimateNoteOrChord = NULL;
     }
     Syl *m_currentSyl;
-    Note *m_lastNote;
-    Note *m_lastButOneNote;
+    LayerElement *m_lastNoteOrChord;
+    LayerElement *m_penultimateNoteOrChord;
 };
 
 //----------------------------------------------------------------------------