+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [unreleased]
+* Support for `beatRpt@beatdef`
+
+## [3.8.1] - 2022-01-10
+* Fix bug in PAE importer for durations and for key signatures
 
 ## [3.8.0] - 2021-12-22
 * Support for figured-bass extenders on `f` without `@tstamp` or `@startid`